+++ core/database/src/main/java/com/into/websoso/core/database/datasource/library/entity/InDatabaseNovelEntity.kt
@@ -1,10 +1,10 @@
-package com.into.websoso.core.database.entity
+package com.into.websoso.core.database.datasource.library.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "novels")
-data class InDatabaseNovelEntity(
+internal data class InDatabaseNovelEntity(
     @PrimaryKey val userNovelId: Long,
     val novelId: Long,
     val title: String,
@@ -20,3 +20,4 @@ data class InDatabaseNovelEntity(
     val keywords: List<String>,
     val myFeeds: List<String>,
 )
+
