+++ data/library/src/main/java/com/into/websoso/data/library/model/NovelEntity.kt
@@ -1,7 +1,5 @@
 package com.into.websoso.data.library.model
 
-import com.into.websoso.core.database.entity.InDatabaseNovelEntity
-
 data class UserNovelsEntity(
     val isLoadable: Boolean,
     val userNovelCount: Long,
@@ -22,39 +20,4 @@ data class NovelEntity(
     val endDate: String,
     val keywords: List<String>,
     val myFeeds: List<String>,
-) {
-    internal fun toNovelDatabase(index: Int): InDatabaseNovelEntity =
-        InDatabaseNovelEntity(
-            userNovelId = userNovelId,
-            novelId = novelId,
-            title = title,
-            novelImage = novelImage,
-            novelRating = novelRating,
-            readStatus = readStatus,
-            isInterest = isInterest,
-            userNovelRating = userNovelRating,
-            attractivePoints = attractivePoints,
-            startDate = startDate,
-            endDate = endDate,
-            keywords = keywords,
-            myFeeds = myFeeds,
-            sortIndex = index,
-        )
-}
-
-fun InDatabaseNovelEntity.toData(): NovelEntity =
-    NovelEntity(
-        userNovelId = userNovelId,
-        novelId = novelId,
-        title = title,
-        novelImage = novelImage,
-        novelRating = novelRating,
-        readStatus = readStatus,
-        isInterest = isInterest,
-        userNovelRating = userNovelRating,
-        attractivePoints = attractivePoints,
-        startDate = startDate,
-        endDate = endDate,
-        keywords = keywords,
-        myFeeds = myFeeds,
-    )
+)
